TNLS-Relayers/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
anchorpy==0.20.1
anchorpy-core==0.2.0
annotated-types==0.7.0
anyio==3.7.1
astroid==2.15.6
asttokens==2.4.1
async-timeout==4.0.3
atomicwrites==1.4.0
attrdict==2.0.1
attrs==20.3.0
autoflake==2.3.1
base58==2.1.1
based58==0.1.1
bech32==1.2.0
betterproto==2.0.0b5
bip32utils==0.3.post4
bitarray==2.8.3
black==23.3.0
blinker==1.8.2
blockdiag==3.0.0
boltons==21.0.0
borsh==0.0.1
borsh-construct==0.1.0
cachetools==5.3.1
certifi==2024.6.2
cffi==1.15.1
charset-normalizer==3.1.0
ckzg==2.0.1
click==8.1.4
colorama==0.4.6
construct==2.10.68
construct-typing==0.5.3
cryptography==40.0.2
cytoolz==0.12.0
decorator==5.1.1
dill==0.3.6
ecdsa==0.16.1
eth-account==0.13.4
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==2.1.0
eth-typing==5.0.0
eth-utils==5.0.0
eth_abi==5.1.0
exceptiongroup==1.2.1
executing==2.0.1
Flask==3.0.0
frozenlist==1.4.0
funcparserlib==2.0.0a0
furl==2.1.3
genpy==2022.1
graphviz==0.20.3
grpclib==0.4.7
h11==0.12.0
h2==4.1.0
hexbytes==1.2.1
hpack==4.0.0
httpcore==0.15.0
httplib2==0.22.0
httpx==0.24.0
hyperframe==6.0.1
idna==3.4
importlib_metadata==7.1.0
iniconfig==2.0.0
ipfshttpclient==0.8.0a2
ipython==8.18.1
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jsonalias==0.1.1
jsonschema==3.2.0
jsonschema-specifications==2023.12.1
lazy-object-proxy==1.10.0
lru-dict==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
miscreant==0.3.0
mnemonic==0.19
more-itertools==8.14.0
multiaddr==0.0.9
multidict==6.0.4
mypy==1.2.0
mypy-extensions==1.0.0
nest-asyncio==1.5.6
netaddr==1.3.0
numpy==2.0.0
orderedmultidict==1.0.1
packaging==24.1
parsimonious==0.10.0
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==10.4.0
plantuml==0.3.0
platformdirs==2.6.2
pluggy==1.5.0
prompt_toolkit==3.0.47
protobuf==3.19.5
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycparser==2.21
pycryptodome==3.18.0
pydantic==2.9.2
pydantic_core==2.23.4
pyflakes==3.2.0
Pygments==2.18.0
pyheck==0.1.5
pylint==2.16.1
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.19.2
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-xprocess==1.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytools==2024.1.5
pyunormalize==15.1.0
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.30.0
rich==13.7.1
rlp==4.0.1
rpds-py==0.18.1
secret-sdk==1.7.1
seqdiag==3.0.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
solana==0.34.2
solders==0.21.0
stack-data==0.6.3
sumtypes==0.1a6
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.5
toolz==0.11.2
traitlets==5.14.3
typer==0.12.3
types-cachetools==5.3.0.7
types-requests==2.32.0.20240914
typing_extensions==4.12.2
urllib3==2.2.1
varint==1.0.2
waitress==3.0.0
wcwidth==0.2.13
web3==7.3.0
webcolors==24.6.0
websockets==13.1
Werkzeug==3.0.3
wrapt==1.16.0
yarl==1.9.4
zipp==3.19.2
